exynos/ipp: add simple API to expose IPPv2 functionality
This wraps the IPP as a context object.

Similar to the G2D a general IPP context has to be created.
From this context we can then query available IPP modules,
create tasks for them, configure the tasks and commit them.
Committing serialises the tasks and passes it to the
kernel driver.